Refactoring around storing data from requests.
Changed StreamHelpers to static class.
Changed names of multiple classes.

+++ TwitchViewerCounter.Core/Models/FeaturedStreamInfo.cs
@@ -5,7 +5,7 @@ namespace TwitchViewerCounter.Core.Models
     public class FeaturedStreamInfo
     {
         [DeserializeAs(Name = "priority")]
-        public long Priority { get; set; }
+        public int Priority { get; set; }
 
         [DeserializeAs(Name = "sponsored")]
         public bool Sponsored { get; set; }
